requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch
transformers
faiss-cpu
Pillow
python-multipart

